Fix up the Makefile a bit
1) `upload` target used to depend on all the files existing in `dist/`,
   which would be no files at all if someone just ran clean. Or some
random files if someone had put some random files in that folder.
Instead, now we just rely on the tarball generated by the script. The
`upload` target uploads those files *specifically*, rather than having a
wildcard for the whole folder.

2) `clean` target now removes `node_modules/`, and `node_modules/` is a
   target of the site tarball. This will make it so that all the targets
will work from a bare repo.

3) I added a target for the `missing_entities` script I wrote in
a419557. This script goes through and looks for nodes that have been
referred to in readings or their subgraphs that don't have any
statements with them as the subject. This is useful for writing new
entries by hand and leaving "stubs" for things like authors, journals,
etc. It's not as good as having some sort of template or form, but it's
fine.