Add a shape for academic articles
I've required the date to be attached to the issue, not to the article
itself. I will make the changes in the commit that immediately follows
this one. Is that a good constraint? In the general case, yes. But there
are some weird articles in journals which don't have issues, in which
case, those should probably be a different shape.